Extract `Matcher` into its own top-level class
